Rename albums BLoC (`.bloc.ts` convention)
By convention now it's `album.bloc.ts`
