+++ web/src/routes/albums/+page.svelte
@@ -9,7 +9,7 @@
 	import NavigationBar from '$lib/components/shared-components/navigation-bar.svelte';
 	import SideBar from '$lib/components/shared-components/side-bar/side-bar.svelte';
 	import PlusBoxOutline from 'svelte-material-icons/PlusBoxOutline.svelte';
-	import { useAlbums } from './albums-bloc';
+	import { useAlbums } from './albums.bloc';
 
 	export let data: PageData;
